:recycle: refactor: Adicionando a extensão dos arquivos .ts e .d.ts em todos as importações.
