+++ ws-clerk-blog/src/repositories/iuser-repository.d.ts
@@ -1,5 +1,5 @@
-import { IUser as IUserCreated } from "../models/iuser-created-event";
-import { IUser as IUserUpdated } from "../models/iuser-updated-event";
+import type { IUser as IUserCreated } from "../models/iuser-created-event.d.ts";
+import type { IUser as IUserUpdated } from "../models/iuser-updated-event.d.ts";
 
 export interface IUserRepository {
   create(user: IUserCreated): Promise<void>;
