+++ ws-clerk-blog/src/services/iclerk-webhook.d.ts
@@ -1,4 +1,4 @@
-import { Request } from "express";
+import type { Request } from "express";
 
 export interface IClerkWeebhookService {
   verify<T>(req: Request): Promise<T | undefined>;
